+++ src/reporter/cucumber/messagesBuilder/TestCase.ts
@@ -12,10 +12,10 @@
 import * as messages from '@cucumber/messages';
 import { TestCaseRun } from './TestCaseRun';
 import { Hook, HookType } from './Hook';
-import { GherkinDocumentWithPickles, PickleWithLocation } from '../../../features/load';
-import { stringifyLocation } from '../../../utils';
+import { GherkinDocumentWithPickles, PickleWithLocation } from '#features/load.js';
+import { stringifyLocation } from '#utils/index.js';
 import { ProjectInfo } from './Projects';
-import { BddData } from '../../../run/bddAnnotation/types';
+import { BddData } from '#run/bddAnnotation/types.js';
 
 type HookWithStep = {
   hook: Hook;