+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Dev
 * update @cucumber/html-formatter ([#213](https://github.com/vitalets/playwright-bdd/issues/213))
+* chore: use import subpath from package.json for internal imports
 
 ## 7.2.2
 * fix: provide empty world for Playwright-style steps ([#208](https://github.com/vitalets/playwright-bdd/issues/208))